refactor: 유효성 검사 수정
- 유저 가입 시 DTO 유효성 검사를 수정하였습니다.
