+++ src/main/java/com/project/growfit/domain/auth/service/impl/AuthParentServiceImpl.java
@@ -107,6 +107,6 @@ public class AuthParentServiceImpl implements AuthParentService {
                 ROLE.fromString("CHILD"));
     }
     private static void updateNickname(AuthParentRequestDto request, Parent parent) {
-        parent.updateNickname(request.nickname());
+        parent.updateInfo(request.nickname());
     }
 }
